Haptic: driver-> feature (#13713)
